Create fasta_qual-select_short-and_fastq.py
Fasta + Qual to FastQ merger no matter sequence or quality length differences. For each pair, it trims the sequences or the quality according to the shortest one.

Only use this script if you are sure that differences are only in 3' UTR and/or if you really know what are you doing.